refactor(eslint-plugin-i18n): rename multiline-i18n-meta-object to multiline-meta
