+++ example/eslint.config.js
@@ -29,7 +29,7 @@ export default [
                     },
                 },
             ],
-            'i18n-eslint-plugin/multiline-i18n-meta-object': 'error',
+            'i18n-eslint-plugin/multiline-meta': 'error',
             'i18n-eslint-plugin/sort-message-locales': 'error',
             'i18n-eslint-plugin/string-literal-keys': [
                 'error',
