+++ packages/eslint-plugin-i18n/rules/multiline-meta/index.ts
